Move staffing rows creation in the staffing service
