+++ src/services/Staffing.js
@@ -2,6 +2,13 @@
 
 import moment from 'moment';
 
+import {
+  calculateTaskWidth,
+  calculateTaskOffsets,
+  calculateWeeklyTasks,
+} from './Task';
+
+
 /**
  * Generate a list of 20 weeks, from four weeks ago to 16 weeks in future.
  *
@@ -19,3 +26,36 @@ export function createWeeks(): Array<moment> {
 
   return weeks;
 }
+
+export function createRows(persons, timeline, weeks) {
+  const rows = [];
+  persons.forEach((person) => {
+    const userTimeline = timeline.filter(task => task.userId === person.id);
+    let weeklyTasksCount = {};
+    let maxWeeklyTasksCount = 0;
+
+    const tasks = userTimeline.map((timelineTask) => {
+      const { xoffset, yoffset } = calculateTaskOffsets(timelineTask, weeklyTasksCount, weeks[0]);
+
+      weeklyTasksCount = calculateWeeklyTasks(timelineTask, weeklyTasksCount);
+      return {
+        timelineTask,
+        xoffset,
+        yoffset,
+        width: calculateTaskWidth(timelineTask),
+      };
+    }, this);
+
+    maxWeeklyTasksCount = Math.max(...Object.values(weeklyTasksCount));
+
+    rows.push({
+      person,
+      tasks,
+      maxWeeklyTasksCount,
+      weeklyTasksCount,
+    });
+  });
+
+  return rows;
+}
+
